add lib/processing/shader.rb
