+++ lib/processing.rb
@@ -13,4 +13,5 @@ require 'processing/vector'
 require 'processing/image'
 require 'processing/font'
 require 'processing/touch'
+require 'processing/shader'
 require 'processing/processing'
